Componentnize NewNoteButton
